+++ projects/qgrid-core/row-details/row.details.let.js
@@ -1,64 +1,20 @@
-import { Command } from '../command/command';
 import { RowDetails } from './row.details';
-import { takeOnce, filter } from '../rx/rx.operators';
-import { toggleStatus, invalidateStatus } from './row.details.service';
-import { selectRow } from '../navigation/navigation.state.selector';
+import { invalidateStatus } from './row.details.service';
+import { RowDetailsToggleStatusCommand } from '../command-bag/row.details.toggle.status.command';
 
 export class RowDetailsLet {
-	constructor(plugin, shortcut) {
-		const { model, observeReply, observe, disposable } = plugin;
+	constructor(plugin) {
+		const { model, observeReply, disposable, commandPalette } = plugin;
 
 		this.plugin = plugin;
 
-		this.toggleStatus = new Command({
-			source: 'row.details.view',
-			execute: row => {
-				if (!row) {
-					row = selectRow(model.navigation());
-				}
-
-				const { toggle, status, mode } = model.row();
-				if (toggle.execute({ row }) !== false) {
-					const newStatus = toggleStatus([row], status, mode);
-					model.row({ status: newStatus }, {
-						source: 'row.details.view'
-					});
-
-					observe(model.sceneChanged)
-						.pipe(
-							filter(e => e.hasChanges('status') && e.state.status === 'stop'),
-							takeOnce()
-						)
-						.subscribe(e => {
-							const rowStatus = newStatus.get(row);
-							if (rowStatus && rowStatus.expand) {
-								const index = model.view().rows.indexOf(row)
-								model.focus({
-									rowIndex: index + 1
-								}, {
-									source: 'row.details.let'
-								});
-							}
-						});
-				}
-			},
-			canExecute: row => {
-				if (!row) {
-					const { cell } = model.navigation();
-					if (cell && cell.column.type === 'row-expand') {
-						row = cell.row;
-					}
-				}
-
-				const { toggle } = model.row();
-				return !!row && toggle.canExecute({ row });
-			},
-			shortcut: model.row().shortcut.toggle
-		});
+		this.toggleStatus = new RowDetailsToggleStatusCommand(plugin);
+		commandPalette.register(this.toggleStatus);
 
 		observeReply(model.sceneChanged)
 			.subscribe(e => {
-				if (e.tag.source === 'row.details.view') {
+				if (e.tag.source === 'row.details.let' ||
+					e.tag.source === 'row.details.toggle.status.command') {
 					return;
 				}
 
@@ -71,8 +27,10 @@ export class RowDetailsLet {
 							mode
 						);
 
-					model.row({ status: newStatus }, {
-						source: 'row.details.view'
+					model.row({
+						status: newStatus
+					}, {
+						source: 'row.details.let'
 					});
 				}
 			});
@@ -100,8 +58,6 @@ export class RowDetailsLet {
 						});
 				}
 			});
-
-		shortcut.register([this.toggleStatus]);
 	}
 
 	status(row) {
@@ -111,6 +67,7 @@ export class RowDetailsLet {
 
 		const { model } = this.plugin;
 		const { status } = model.row();
+
 		const state = status.get(row);
 		return state && state.expand ? 'expand' : 'collapse';
 	}