+++ CHANGELOG.md
@@ -4,9 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [9.2.0] - 2020-06-02
+### BREAKING
+- Rename $view.filter.column to $view.filter.row.
+- Remove `options` property from $view.edit.cell.
+
 ### Added
 - Support 2 version 9.1.x - Ivy, 9.1.x-ve - View Engine
 - `q-grid-shortcut-host` directive that can change key event source for q-grid commands.
+- Command palette API to access grid core commands.
+
+### Changed
+- Shortcut system.
 
 ### Changed
 - Better top toolbar layout.